Create README.md
This is a segment of my project that i've made for integrating in Loan Approval Prediction Project.

Here Opencv Technology is Used in which ORB is used and Bruete Force Matcher is used for Matching descriptor of between two images
and warping perspective image and used image masking to crop image in two parts which is Full Name and PanCard Number.

Then Tesseract Functionality is used which is OCR Technology to extract text from image.

so 2 cropped images's text are extracted and stored in list.

then on frontend input parameter,this parameter go in backend and checking that whether it matched with uploaded document and input parameter.

here we have used this technology for reference because we can't have gov. api access of document verifying for Just Personal Project.